k-distribution/tests/pyk: add simple-proofs testing harness
